LPD-15185 Rename pages to follow the new patter
