+++ modules/test/playwright/pages/product-navigation-applications-menu/AICreatorSettingsPage.ts
